+++ taglib/dsf/dsfproperties.cpp
@@ -129,6 +129,6 @@ void DSF::Properties::read(const ByteVector &data)
   d->bitrate = static_cast<unsigned int>(
       d->samplingFrequency * d->bitsPerSample * d->channelNum / 1000.0 + 0.5);
   d->length = d->samplingFrequency > 0
-      ? static_cast<unsigned int>(d->sampleCount * 1000.0 / d->samplingFrequency + 0.5)
+      ? static_cast<unsigned int>(static_cast<double>(d->sampleCount) * 1000.0 / d->samplingFrequency + 0.5)
       : 0;
 }
