+++ taglib/mpc/mpcproperties.cpp
@@ -229,7 +229,7 @@ void MPC::Properties::readSV8(File *file, offset_t streamLength)
          frameCount > 0 && d->sampleRate > 0) {
         const auto length = static_cast<double>(frameCount) * 1000.0 / d->sampleRate;
         d->length  = static_cast<int>(length + 0.5);
-        d->bitrate = static_cast<int>(streamLength * 8.0 / length + 0.5);
+        d->bitrate = static_cast<int>(static_cast<double>(streamLength) * 8.0 / length + 0.5);
       }
     }
     else if (packetType == "RG") {
@@ -331,6 +331,6 @@ void MPC::Properties::readSV7(const ByteVector &data, offset_t streamLength)
     d->length = static_cast<int>(length + 0.5);
 
     if(d->bitrate == 0)
-      d->bitrate = static_cast<int>(streamLength * 8.0 / length + 0.5);
+      d->bitrate = static_cast<int>(static_cast<double>(streamLength) * 8.0 / length + 0.5);
   }
 }
